Restore the get(NDArray, NDIndex)
