+++ api/src/main/java/ai/djl/ndarray/index/NDArrayIndexer.java
@@ -20,6 +20,7 @@ import ai.djl.ndarray.index.full.NDIndexFullPick;
 import ai.djl.ndarray.index.full.NDIndexFullSlice;
 
 import java.util.List;
+import java.util.Optional;
 
 /** A helper class for {@link NDArray} implementations for operations with an {@link NDIndex}. */
 public abstract class NDArrayIndexer {
@@ -49,7 +50,33 @@ public abstract class NDArrayIndexer {
      * @param index the index to get
      * @return the subarray
      */
-    public abstract NDArray get(NDArray array, NDIndex index);
+    public NDArray get(NDArray array, NDIndex index) {
+        if (index.getRank() == 0 && array.getShape().isScalar()) {
+            return array.duplicate();
+        }
+
+        // use booleanMask for NDIndexBooleans case
+        List<NDIndexElement> indices = index.getIndices();
+        if (!indices.isEmpty() && indices.get(0) instanceof NDIndexBooleans) {
+            if (indices.size() != 1) {
+                throw new IllegalArgumentException(
+                        "get() currently doesn't support more that one boolean NDArray");
+            }
+            return array.booleanMask(((NDIndexBooleans) indices.get(0)).getIndex());
+        }
+
+        Optional<NDIndexFullPick> fullPick = NDIndexFullPick.fromIndex(index, array.getShape());
+        if (fullPick.isPresent()) {
+            return get(array, fullPick.get());
+        }
+
+        Optional<NDIndexFullSlice> fullSlice = NDIndexFullSlice.fromIndex(index, array.getShape());
+        if (fullSlice.isPresent()) {
+            return get(array, fullSlice.get());
+        }
+        throw new UnsupportedOperationException(
+                "get() currently supports all, fixed, and slices indices");
+    }
 
     /**
      * Sets the values of the array at the fullSlice with an array.
